Renamed OptionalDoubleShouldContainValueCloseTo to OptionalDoubleShouldHaveValueCloseTo
